data: SmolLM2-135M baseline — 0/48 exact match (0%)
Zero-shot baseline with HuggingFaceTB/SmolLM2-135M (135M params, direct mode):
  Exact Match:      0/48 (0.0%)
  Normalized Match: 1/48 (2.1%)
  BASE_10_ERROR:    0
  TOKEN_COLLAPSE:   1
  FORMAT_HALLUCINATION: 5
  OTHER_ERROR:      41

The model is essentially random on temporal arithmetic.
Validates the measurement pipeline for future Gemma-2B comparison.

Also patches validate.py: float32 → float16 for M-series Mac compatibility.

Addresses: Issue #20 (baseline measurement — partial, SmolLM proxy)

+++ tools/validate.py
@@ -249,7 +249,7 @@ def load_model(model_id, adapter_path=None, backend="auto"):
         import torch
         from transformers import AutoTokenizer, AutoModelForCausalLM
         print(f"Loading {model_id} via PyTorch...")
-        model = AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=torch.float32, device_map="auto")
+        model = AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=torch.float16, device_map="auto")
         tokenizer = AutoTokenizer.from_pretrained(model_id)
         if tokenizer.pad_token is None:
             tokenizer.pad_token = tokenizer.eos_token
